+++ frontend/lib/widgets/navbar/navbar-v2.dart
